编写武器和怪物信息加载; 编写时间类

+++ src/monster.hpp
@@ -17,7 +17,7 @@ extern std::string monsterConfigPath;
 class monsterDrop {
 public:
     LL          id;
-    LL          chance;
+    double      chance;
 };
 
 class monsterData {
